src/functions/osint_collector/requirements.txt
google-cloud-storage
requests
functions-framework
python-dotenv
